Reorganise data related classes
