+++ src/main/java/seedu/address/commons/dataset/date/MonthlyCountData.java
@@ -1,4 +1,4 @@
-package seedu.address.commons;
+package seedu.address.commons.dataset.date;
 
 import static java.util.Objects.requireNonNull;
