+++ src/main/java/seedu/address/logic/commands/meeting/MultipleMeetingStatsCommand.java
@@ -7,7 +7,7 @@ import java.time.Month;
 import java.time.Year;
 import java.time.ZoneId;
 
-import seedu.address.commons.MonthlyCountDataSet;
+import seedu.address.commons.dataset.date.MonthlyCountDataSet;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
