+++ src/main/java/seedu/address/model/ModelManager.java
@@ -14,7 +14,7 @@ import java.util.logging.Logger;
 import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import javafx.collections.transformation.SortedList;
-import seedu.address.commons.MonthlyCountDataSet;
+import seedu.address.commons.dataset.date.MonthlyCountDataSet;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.model.meeting.Meeting;
