+++ src/test/java/seedu/address/commons/MonthlyCountDataTest.java
@@ -10,6 +10,9 @@ import java.time.Year;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.commons.dataset.date.MonthAndYear;
+import seedu.address.commons.dataset.date.MonthlyCountData;
+
 
 public class MonthlyCountDataTest {
