+++ src/test/java/seedu/address/logic/commands/meeting/MultipleMeetingStatsCommandTest.java
@@ -12,7 +12,7 @@ import java.time.ZoneId;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.commons.MonthlyCountDataSet;
+import seedu.address.commons.dataset.date.MonthlyCountDataSet;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
